Desktop/xx/backend/pyproject.toml
[tool.poetry]
name = "core"
version = "0.1.0"
description = ""
authors = ["Ratchaphon1412 <Ratchaphon.h111@gmail.com>"]
readme = "README.md"
packages = [
    { include = "core", from = "." }
]

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.1.6"
djangorestframework = "^3.15.2"
drf-spectacular = "^0.28.0"
drf-spectacular-sidecar = "^2025.2.1"
minio = "^7.2.15"
django-seed = "^0.3.1"
psycopg = "^3.2.4"
django-cors-headers = "^4.7.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
